requirements.txt
Django==2.1.5
django-allauth==0.38.0
django-cors-headers==2.4.0
django-environ==0.4.5
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
psycopg2-binary-2.7.6.1
